Updated fully basic needed Controller

+++ src/main/java/com/shoppingweb/DAO/CategoryRepository.java
@@ -6,7 +6,7 @@ import com.shoppingweb.Service.CategoryService;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.*;
 
-public interface CategoryRepository extends JpaRepository<Category,Integer>{
+public interface CategoryRepository extends JpaRepository<Category,Long>{
     List<Category> findAll();
     Category save(Category category);
     void deleteById(int id);
