+++ src/main/java/com/shoppingweb/Service/CategoryService.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public interface CategoryService {
     public List<Category> getAllCategories();
-    public Category getCategoryById(int theId);
+    public Category getCategoryById(Long theId);
     public void addCategory();
     public void updateCategory();
     public void deleteCategory();
